refactor!: split generator and widget to multiple pages
