+++ index.html
@@ -10,6 +10,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script type="module" src="/src/main.tsx"></script>
+<script type="module" src="/src/App.tsx"></script>
 </body>
 </html>
