+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "faceit-stats-widget",
   "private": true,
-  "version": "2.2.1",
+  "version": "3.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
