+++ src/App.tsx
@@ -1,12 +1,10 @@
-import {Route, Routes} from "react-router-dom";
-import {Widget} from "./widget/Widget.tsx";
+import {BrowserRouter} from "react-router-dom";
 import {Generator} from "./generator/Generator.tsx";
+import './styles/app.less'
+import ReactDOM from "react-dom/client";
 
-export const App = () => {
-  return <>
-    <Routes>
-      <Route path="/widget/generator" element={<Generator/>}/>
-      <Route path="/widget/" element={<Widget/>}/>
-    </Routes>
-  </>
-}
+ReactDOM.createRoot(document.getElementById('root')!).render(
+  <BrowserRouter>
+    <Generator/>
+  </BrowserRouter>,
+)
