+++ vite.config.ts
@@ -4,4 +4,12 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build: {
+    rollupOptions: {
+      input: {
+        main: './index.html',
+        widget: './widget/index.html',
+      }
+    }
+  }
 })
