+++ widget/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+  <title>FACEIT Stats Widget</title>
+</head>
+<body>
+<div id="root"></div>
+<script type="module" src="/widget/src/App.tsx"></script>
+</body>
+</html>
