+++ widget/src/styles/color_schemes.less
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=DM+Sans:ital,opsz,wght@0,9..40,100..1000;1,9..40,100..1000&display=swap');
+
 /* Color scheme mixin */
 .scheme(@text, @subtext, @background, @border1, @border2, @borderRotation, @wins, @losses, @bg-opacity: 0.1) {
   --text: @text;
