+++ widget/src/styles/themes/compact.less
